Extract action menu into its own module
